Rename nsc.backend.jvm.CodeGenTools to testing.BytecodeTesting
