+++ test/junit/scala/issues/OptimizedBytecodeTest.scala
@@ -6,9 +6,9 @@ import org.junit.Test
 import scala.tools.asm.Opcodes._
 import org.junit.Assert._
 
-import scala.tools.nsc.backend.jvm.{AsmUtils, CodeGenTools}
+import scala.tools.nsc.backend.jvm.AsmUtils
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import AsmUtils._
