+++ test/junit/scala/tools/nsc/backend/jvm/DirectCompileTest.scala
@@ -4,7 +4,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 import org.junit.Assert._
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.asm.Opcodes._
 import scala.tools.partest.ASMConverters._
 import scala.tools.testing.ClearAfterClass
