+++ test/junit/scala/tools/nsc/backend/jvm/StringConcatTest.scala
@@ -9,7 +9,7 @@ import org.junit.Assert._
 
 import scala.tools.testing.AssertUtil._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import scala.tools.testing.ClearAfterClass
