+++ test/junit/scala/tools/nsc/backend/jvm/opt/BTypesFromClassfileTest.scala
@@ -11,7 +11,7 @@ import org.junit.Assert._
 import scala.tools.nsc.backend.jvm.BTypes.InternalName
 import scala.tools.testing.AssertUtil._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
