+++ test/junit/scala/tools/nsc/backend/jvm/opt/CallGraphTest.scala
@@ -15,7 +15,7 @@ import scala.tools.asm.tree.analysis._
 import scala.tools.nsc.reporters.StoreReporter
 import scala.tools.testing.AssertUtil._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import AsmUtils._
