+++ test/junit/scala/tools/nsc/backend/jvm/opt/ClosureOptimizerTest.scala
@@ -17,7 +17,7 @@ import scala.tools.nsc.io._
 import scala.tools.nsc.reporters.StoreReporter
 import scala.tools.testing.AssertUtil._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import AsmUtils._
