+++ test/junit/scala/tools/nsc/backend/jvm/opt/InlineInfoTest.scala
@@ -8,7 +8,7 @@ import org.junit.Test
 import scala.collection.generic.Clearable
 import org.junit.Assert._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import AsmUtils._
