+++ test/junit/scala/tools/nsc/backend/jvm/opt/InlinerIllegalAccessTest.scala
@@ -11,7 +11,7 @@ import org.junit.Assert._
 import scala.tools.asm.tree._
 import scala.tools.testing.AssertUtil._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import AsmUtils._
