+++ test/junit/scala/tools/nsc/backend/jvm/opt/InlinerTest.scala
@@ -12,7 +12,7 @@ import org.junit.Assert._
 import scala.tools.asm.tree._
 import scala.tools.nsc.reporters.StoreReporter
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import AsmUtils._
