+++ test/junit/scala/tools/nsc/backend/jvm/opt/SimplifyJumpsTest.scala
@@ -8,7 +8,7 @@ import org.junit.Test
 import scala.tools.asm.Opcodes._
 import org.junit.Assert._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
