+++ test/junit/scala/tools/nsc/backend/jvm/opt/UnreachableCodeTest.scala
@@ -10,7 +10,7 @@ import org.junit.Assert._
 
 import scala.tools.testing.AssertUtil._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import scala.tools.testing.ClearAfterClass
