+++ test/junit/scala/tools/nsc/backend/jvm/opt/UnusedLocalVariablesTest.scala
@@ -9,7 +9,7 @@ import scala.tools.asm.Opcodes._
 import org.junit.Assert._
 import scala.collection.JavaConverters._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import scala.tools.testing.ClearAfterClass
