+++ test/junit/scala/tools/nsc/transform/patmat/PatmatBytecodeTest.scala
@@ -8,10 +8,9 @@ import scala.tools.asm.Opcodes._
 import org.junit.Assert._
 
 import scala.tools.nsc.backend.jvm.AsmUtils._
-import scala.tools.nsc.backend.jvm.CodeGenTools
 import scala.tools.testing.AssertUtil._
 
-import CodeGenTools._
+import scala.tools.testing.BytecodeTesting._
 import scala.tools.partest.ASMConverters
 import ASMConverters._
 import scala.tools.testing.ClearAfterClass
